+++ CHANGELOG.rst
@@ -27,5 +27,6 @@
 0.3.0
 -----
 - Add a test suite
-- Py3 compatibility
+- Py3 compatible
 - Importable modules
+- Supports ascii/fits catalogs